testing : moved the audio gen function from app to utils.py + replace the generation call in generation podcast audio segments function cz  using the generate_long_text_optimized directly  caused some issue while shifting roles in the podcast generation sometimes
